Avoid circular reference in task processing platform_bus
Since the task spawned in session.start() was holding a strong
(circular) reference to the session it wasn't possible to drop
a Session and have it free all associated resources.

This follows a pattern that's also being used in the backend of
passing around a Weak pointer to the Session 'inner' and having
a Session::wrap() api that can be used to get a temporary
Session reference just while processing a backend event.