Add Gatt services + characteristics support
This is a fairly huge code dump really but is generally about
adding support for discovering Gatt services (+ included
services) and characteristics and being able read/write
characteristic values and subscribe to notifications for
characteristics.

The API allows for Peripherals to export the same service
multiple times and tries to preserve the order they have
on-device in case that's significant. Secondary services
can be found through Service includes and here we try and
preserve the order of the includes. Characteristics can
be queried via a Service and again the API allows for
a service to expose multiple instances of the same
characteristic and tries to preserve the on-device order.

This doesn't yet include support for accessing characteristic
descriptors.

This adds a heart-rate-monitor.rs example that demonstrates
using the current API to scan for a heart rate monitor, then
connect and subscribe to heart rate notifications.

This also makes a few sweeping changes including renaming
/PlatformPeripheralHandle/PeripheralHandle/ and renaming
/Platform/Backend/ as well as cargo fmt changes